Move amcrest services to separate module
